Frontend: Finish moving parser contract tests out of subroutine suite
